Refactor yay
