+++ src/lib.rs
@@ -1,3 +1,8 @@
+/// Errors that can be generated.
 pub mod errors;
+/// Utilities to work with Presentations
 pub mod presentation;
+/// UI utilities
+/// The UI accepts input from the command line, and
+/// can also read a config file
 pub mod ui;
